LayoutTests/crypto/subtle/hmac-generate-export-raw-key.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="../resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Test exporting an HMAC key with raw format");

jsTestIsAsync = true;

var extractable = true;

debug("Generating a key...");
crypto.subtle.generateKey({name: "hmac", hash: "sha-1"}, extractable, ["sign", "verify"]).then(function(key) {
    debug("Exporting a key...");
    return crypto.subtle.exportKey("raw", key);
}).then(function(result) {
    key = result;
    shouldBe("key.byteLength", "64");

    finishJSTest();
});
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
