Safari-611.1.10/LayoutTests/crypto/subtle/hmac-generate-export-raw-key-expected.txt

Test exporting an HMAC key with raw format

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Generating a key...
Exporting a key...
PASS key.byteLength is 64
PASS successfullyParsed is true

TEST COMPLETE
